Inline toolbar.tsx into toolbar/index.tsx
